+++ Package.swift
@@ -16,9 +16,7 @@ let package = Package(
             targets: ["StoreKitPlus"]
         ),
     ],
-    dependencies: [
-        .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.0.0"),
-    ],
+    dependencies: [],
     targets: [
         .target(
             name: "StoreKitPlus",