+++ Sources/StoreKitPlus/Products/ProductID.swift
@@ -1,0 +1,15 @@
+//
+//  ProductID.swift
+//  StoreKitPlus
+//
+//  Created by Daniel Saidi on 2022-08-24.
+//  Copyright © 2022 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/**
+ This typealias defines the type used for product IDs in the
+ StoreKit and StoreKitPlus packages.
+ */
+public typealias ProductID = String